% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1cent.R
\name{Extract}
\alias{Extract}
\alias{[.L1cent}
\alias{`[.L1cent`}
\title{Extracting in the L1centrality Package}
\usage{
\method{[}{L1cent}(x, i)
}
\arguments{
\item{x}{An object used to select a method.}

\item{i}{Indices specifying elements to extract.}
}
\value{
Extract parts of an object.
}
\description{
Subsetting method for the classes in the \code{L1centrality} package. See details below.
}
\seealso{
\code{\link[=L1cent]{L1cent()}} returns an object of class \code{L1cent}.
}
\keyword{internal}
