% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1cent.R
\name{Summary}
\alias{Summary}
\alias{summary.L1cent}
\title{Summarizing the L1centrality Package}
\usage{
\method{summary}{L1cent}(object, ...)
}
\arguments{
\item{object}{An object used to select a method.}

\item{...}{Further arguments passed to or from other methods. This argument
is ignored here.}
}
\value{
A numeric of length seven, consisting of the five-number summary,
mean, and the Gini coefficient.
}
\description{
\code{summary()} method for the classes in the \code{L1centrality} package. See details below.
}
\seealso{
\code{\link[=L1cent]{L1cent()}} returns an object of class \code{L1cent}.

\code{\link[=Lorenz_plot]{Lorenz_plot()}} computes the Gini coefficient as well.
}
\keyword{internal}
