% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1centrality-package.R
\docType{package}
\name{L1centrality-package}
\alias{L1centrality-package}
\alias{_PACKAGE}
\alias{L1centrality}
\title{L1centrality: Graph/Network Analysis based on \out{<i>L</i><sub>1</sub>}
Centrality}
\description{
Analyze graph/network data using \out{<i>L</i><sub>1</sub>} centrality.
Functions for deriving global and local \out{<i>L</i><sub>1</sub>}
centralities and \out{<i>L</i><sub>1</sub>} centrality-based neighborhoods of
vertices are provided. Routines for visual inspection of a graph/network are
also provided.
}
\references{
S. Kang and H.-S. Oh. On a notion of graph centrality based on
\ifelse{html}{\out{<i>L</i><sub>1</sub>}}{{\eqn{L_1}}} data depth.
Technical Report, 2023.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seungwoo-stat/L1centrality}
  \item Report bugs at \url{https://github.com/seungwoo-stat/L1centrality/issues}
}

}
\author{
\strong{Maintainer}: Seungwoo Kang \email{kangsw0401@snu.ac.kr} (\href{https://orcid.org/0000-0001-8082-0794}{ORCID})

Authors:
\itemize{
  \item Hee-Seok Oh \email{heeseok@stats.snu.ac.kr}
}

}
\keyword{internal}
